move show_mem to driver
