+++ Project.toml
@@ -4,7 +4,6 @@ authors = ["CliMA Contributors <clima-software@caltech.edu>"]
 version = "0.1.0"
 
 [deps]
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 ClimaAtmos = "b2c96348-7fb7-4fe0-8da9-78d88439e717"
 ClimaComms = "3a4d1b5c-c61d-41fd-a00a-5873ba7a1b0d"
 ClimaCore = "d414da3d-4745-48bb-8d80-42e94e092884"
@@ -25,7 +24,6 @@ TempestRemap_jll = "8573a8c5-1df0-515e-a024-abad257ee284"
 Thermodynamics = "b60c26fb-14c3-4610-9d3e-2d17fe7ff00c"
 
 [compat]
-CUDA = "5.2"
 ClimaAtmos = "0.22"
 ClimaComms = "0.5.6"
 ClimaCore = "0.13"
