+++ experiments/AMIP/coupler_driver.jl
@@ -25,6 +25,35 @@ so we force abbreviated stacktraces even in non-interactive runs.
 =#
 using CUDA
 
+"""
+    show_memory_usage(comms_ctx, objects)
+
+Display the current memory footprint of the simulation, using an appropriate
+method based on the device being used.
+
+In the GPU case, show the memory usage of the GPU.
+In the CPU case, show the memory footprint of the provided object(s).
+Note that these two cases provide different information, and should not be
+directly compared.
+
+# Arguments
+`comms_ctx`: the communication context being used to run the model
+`objects`: Dict mapping objects whose memory footprint is displayed in the CPU case to their names
+"""
+function show_memory_usage(comms_ctx, objects)
+    if comms_ctx.device isa ClimaComms.CUDADevice
+        @info "Memory usage: $(CUDA.memory_status())"
+    elseif comms_ctx.device isa ClimaComms.AbstractCPUDevice
+        if ClimaComms.iamroot(comms_ctx)
+            for (obj, name) in objects
+                @info "Memory footprint of `$(name)` in bytes: $(Base.summarysize(obj))"
+            end
+        end
+    else
+        @warn "Invalid device type $device; cannot show memory usage."
+    end
+end
+
 redirect_stderr(IOContext(stderr, :stacktrace_types_limited => Ref(false)))
 
 #=
@@ -65,7 +94,7 @@ using ClimaCoupler.Regridder
 using ClimaCoupler.Regridder: update_surface_fractions!, combine_surfaces!, binary_mask
 using ClimaCoupler.TimeManager:
     current_date, Monthly, EveryTimestep, HourlyCallback, MonthlyCallback, update_firstdayofmonth!, trigger_callback!
-import ClimaCoupler.Utilities: get_comms_context, show_memory_usage
+import ClimaCoupler.Utilities: get_comms_context
 
 pkg_dir = pkgdir(ClimaCoupler)
